pyproject.toml
[tool.poetry]
name = "asus-charge-daemon"
version = "0.3.1"
description = "A daemon exposing D-Bus services for managing the charge level of recent ASUS notebooks."
authors = ["Christopher Forrester <christopher@cforrester.ca>"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com./cforrester1988/asus-charge-daemon"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: No Input/Output (Daemon)",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Topic :: System",
]
packages = [{ include = "asuscharged" }]
include = ["LICENCE", "install.py", "data/*"]

[tool.poetry.scripts]
asuscharged = "asuscharged.__main__:main"

[tool.poetry.dependencies]
python = "^3.7"
asus-charge-control = "^1.0.3"
dbus-next = "^0.2.2"
asyncinotify = "^2.0.2"
PyGObject = "^3.38.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^20.8b1"
pylama = "^7.7.1"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
